backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy>=2.0.36
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0
bcrypt==4.1.2
MetaTrader5>=5.0.4510
cryptography==42.0.0
redis==5.0.1
celery==5.3.6
python-dotenv==1.0.0
aiofiles==23.2.1
pillow==10.2.0
httpx==0.26.0
slowapi==0.1.9
